refactor(cli): extract runtime preflight and install helpers from start.ts
Pulls scattered preflight logic out of start.ts into two focused modules:

- lib/install-helpers.ts: shared install primitives (askYesNo,
  runInteractiveCommand, tryInstallWithAttempts, genericInstallHints,
  canPromptForInstall, InstallAttempt) — used by ensureGit/ensureTmux,
  the agent runtime installer, and the optional gh install path.

- lib/startup-preflight.ts: the runtime checks themselves (ensureGit,
  ensureTmux, warnAboutLegacyStorage, warnAboutOpenClawStatus) plus a
  top-level runtimePreflight(config) that orchestrates tools + state
  warnings + idle-sleep + OpenClaw credentials. Distinct from the
  existing lib/preflight.ts, which validates dashboard build artifacts.

start.ts now calls runtimePreflight(config) once at the top of
runStartup instead of inlining 32 lines of orchestration; ensureGit is
imported for the three callsites that still need it directly (URL
clone, addProjectToConfig, URL-while-running branch).

No behavior change. Test count and failure set are identical to
upstream/main (8 pre-existing failures in the stop command tests).

Step 1 of PR A in the start.ts refactor plan
(~/.ao/agent-orchestrator/ao-118/plan.md). start.ts shrinks by 301
lines; net +55 LOC across the touched files (the abstractions cost
some interface overhead, as expected).
